services/payments/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
stripe==10.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
